Update on Web Responsive (I barely survived, but it was enjoyable nonetheless))
*Mobile-first  responsive website
*Managed to  somehow align items, however I seem to encounter a whole lot of problems in the:
>lg-xl web size, menu list *contact us* is not arranged properly, tried to adjust it by adding column number but it just go down to another row.. hmm..
>md-xl, I wanted to adjust the size of the slide images in the carousel however, I haven't figured out how to go about it.. tried using rows and columns, and adjusting width of image by the class w-{value}.
>xs-lg, cards sizing, unfortunately, I am finding it quite tricky to give equal sizes for the cards.. lolz.
>md-xl still I can't seem to properly align to center items in the review section. justify-content-center and align-items-center are not working or i might  have just placed it in the wrong line, but I already tried placing those classes in different lines within the section, however, no effect is being done.

Sentiment: I barely survived, but it was enjoyable, nonetheless -- figuring out how to go about the adjustments for a web responsive page. I realized how important the rows and columns in making a responsive website. haha! I further read in the bootstrap documentation that they will be upgrading it or maybe just adding a class feature "grid" as an alternative for web responsiveness.

Please do not hesitate to give the worst and the best feedback for the web page.

Thank you!